+++ src/PSRule.Rules.Azure/rules/Azure.Cosmos.Rule.ps1
@@ -17,6 +17,26 @@ Rule 'Azure.Cosmos.DefenderCloud' -Ref 'AZR-000382' -Type 'Microsoft.DocumentDb/
 Rule 'Azure.Cosmos.DisableLocalAuth' -Ref 'AZR-000420' -Type 'Microsoft.DocumentDb/databaseAccounts' -If { Test-IsNoSQL } -Tag @{ release = 'GA'; ruleSet = '2024_06'; 'Azure.WAF/pillar' = 'Security'; } -Labels @{ 'Azure.MCSB.v1/control' = 'IM-1'; 'Azure.WAF/maturity' = 'L1' } {
     $Assert.HasFieldValue($TargetObject, 'properties.DisableLocalAuth', $true)
 }
+
+# Synopsis: Use zone redundant Cosmos DB accounts in supported regions to improve reliability.
+Rule 'Azure.Cosmos.AvailabilityZone' -Ref 'AZR-000502' -Type 'Microsoft.DocumentDb/databaseAccounts' -Tag @{ release = 'GA'; ruleSet = '2025_12'; 'Azure.WAF/pillar' = 'Reliability'; } -Labels @{ 'Azure.WAF/maturity' = 'L1' } {
+    # Check for availability zones based on virtual machine scale sets, because it is not exposed through the provider for Cosmos DB.
+    $provider = [PSRule.Rules.Azure.Runtime.Helper]::GetResourceType('Microsoft.Compute', 'virtualMachineScaleSets');
+
+    $Assert.GreaterOrEqual($TargetObject, 'properties.locations', 1);
+
+    if ($TargetObject.properties.locations) {
+        foreach ($location in $TargetObject.properties.locations) {
+            $availabilityZones = GetAvailabilityZone -Location $location.locationName -Zone $provider.ZoneMappings;
+
+            # If the location supports availability zones, ensure zone redundancy is enabled
+            if ($availabilityZones) {
+                $Assert.HasFieldValue($location, 'isZoneRedundant', $true).
+                    ReasonFrom('properties.locations', $LocalizedData.CosmosDBAvailabilityZone, $TargetObject.Name, $location.locationName);
+            }
+        }
+    }
+}
 #endregion Rules
 
 #region Helper functions